A file can say anything, including about itself
The other half of what crosses the boundary is bytes: a picture, a
drawing, a face, a colour profile, a document. Every one is a file
somebody chose, so every one can be nothing, noise, or the first half
of something real — which is what a download that stopped looks like.
Each is now handed each of those, and the page drawn afterwards.

A .chitra gets its own, because it is the one that can lie. The page's
size and every resource's come out of its manifest while the pixels
come out of entries beside it, so those sizes are numbers arriving from
outside exactly as the app's are. The page has been held to what can be
drawn since early on. A resource had not: a file naming one of 65536 by
65536 overflowed the very check meant to hold its bytes against it —
a panic in a debug build, and in a release one a wrap that lets sixteen
bytes stand for an enormous picture. Counted in a width that holds it
now.

+++ core/doc/src/lib.rs
@@ -285,7 +285,14 @@ impl Document {
     /// deserialized manifest (bytes are stored outside the manifest).
     pub fn restore_resource_bytes(&mut self, id: &str, rgba8: Vec<u8>) -> bool {
         match self.resources.get_mut(id) {
-            Some(r) if rgba8.len() == (r.width * r.height * 4) as usize => {
+            // Counted in a width that holds it: the sizes come out of a
+            // file's manifest and the bytes out of the file beside it,
+            // so this is the one place they are made to agree — and a
+            // pair whose pixels do not fit a `u32` would have failed
+            // here as arithmetic rather than as a mismatch, which in a
+            // release build is worse: the product wraps, and a handful
+            // of bytes can be accepted for an enormous picture.
+            Some(r) if rgba8.len() as u64 == r.width as u64 * r.height as u64 * 4 => {
                 r.rgba8 = rgba8;
                 true
             }
